pyproject.toml
[tool.poetry]
name = "hacky_cloud_drive"
version = "0.1.0"
description = ""
authors = ["Michał Freliszka <mfreliszka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.6"
djangorestframework = "3.15.2"
ruff = "^0.9.9"
uvicorn = "^0.34.0"
djangorestframework-simplejwt = "^5.5.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
drf-spectacular = "^0.28.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
